requirements.txt
Django
djangorestframework
Pillow
python-magic
django-filter
django-cors-headers
psutil
redis
django-prometheus
django-health-check
django-debug-toolbar
